add sass scaffolding
