+++ package.json
@@ -4,6 +4,7 @@
   "dependencies": {
     "font-awesome": "^4.7.0",
     "jquery": "^3.4.0",
+    "node-sass": "^4.11.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-fittext": "^1.0.0",
