+++ src/Components/About/About.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import './About.scss'
 
 class About extends Component {
   render() {
