+++ src/Components/About/About.scss
@@ -1,0 +1,1 @@
+// Add The About SASS here
