+++ src/Components/Footer/Footer.scss
@@ -1,0 +1,1 @@
+// Add some SASS
