+++ src/Components/Header/Header.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import ReactFitText from 'react-fittext'
+import './Header.scss'
 
 class Header extends Component {
   render() {
@@ -20,7 +21,7 @@ class Header extends Component {
         <ul id="nav" className="nav">
           <li className="current"><a className="smoothscroll" href="#home">Home</a></li>
           <li><a className="smoothscroll" href="#about">About</a></li>
-          <li><a className="smoothscroll" href="#project">Works</a></li>
+          <li><a className="smoothscroll" href="#project">Projects</a></li>
           <li><a className="smoothscroll" href="#testimonials">Testimonials</a></li>
         </ul>
