+++ src/Components/Header/Header.scss
@@ -1,0 +1,1 @@
+// Migrate all header styles
