+++ src/Components/Testimonials/Testimonials.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import './Testimonials.scss'
 
 class Testimonials extends Component {
   render() {
