+++ src/css/layout.css
@@ -5,6 +5,8 @@
 
 body { background: #0f0f0f; }
 
+#home {background-image: url(../images/bg.png)}
+
 /* ------------------------------------------------------------------ */
 /* b. Header Styles
 /* ------------------------------------------------------------------ */
@@ -921,7 +923,7 @@ footer {
 }
 
 footer a, footer a:visited { color: #525252; }
-footer a:hover, footer a:focus { color: #fff; }
+/* footer a:hover, footer a:focus { color: #fff; }  TODO: The link disappears, make it better*/
 
 /* copyright */
 footer .copyright {
